Update index.html: do not set the publish date in the editor's draft
This is an answer to #196 

What is "not being updated" is the editor's draft; the official draft on the W3C site is o.k. The reason is that the editor's draft includes an explicit publication date (and it should not, it should reflect the latest changes). This PR fixes that.